requirements.txt
altgraph==0.17.4
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
et_xmlfile==2.0.0
FreeSimpleGUI==5.2.0.post1
h11==0.16.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySimpleGUI==5.0.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
rsa==4.9.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
